backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
django-cors-headers
python-dotenv
requests
gitpython
PyGithub
psycopg2-binary
celery
redis
aiohttp
